% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetIm.R
\name{MetIm}
\alias{MetIm}
\title{Impute missing methylation data in sequence and microarray to their union dimensions}
\usage{
MetIm(sequence, microarray = NULL, lambda = NULL, use.mvalue = NULL,
  cvfold = NULL, partition = NULL, cut = NULL, qc_frac = NULL,
  overlap = NULL, ...)
}
\arguments{
\item{sequence}{is the sequencing-based methylation measurement, usually only for a small number of individuals but almost all CPG sites}

\item{microarray}{is the array-based methylation measurement, usually only for a small number of CpG sites, but almost all individuals.}

\item{lambda}{is a parameter controlling for the shrinkage}

\item{use.mvalue}{is a paramter determine the option if MVN will be calculated in the M-value scale. Default=TRUE}

\item{cvfold}{is a fold times of cross validation.}

\item{partition}{is partition.}

\item{cut}{is cut.}

\item{qc_frac}{is qc_frac}

\item{overlap}{is overlap.}
}
\value{
a MetIM object, (not fully implemented yet)
}
\description{
Impute missing methylation data in sequence and microarray to their union dimensions
}
\examples{
data(dat1)
data(dat2)
qc_frac = 1-5/dim(dat2)[2]
Seq <- MetIm(sequence = t(dat1), microarray = t(dat2), lambda=0.3, cut=10, cvfold=0, use.mvalue = T, qc_frac = qc_frac)
}

